Update and bugs fixed.  refactory the project

+++ ChlauApp/Board/Board.py
@@ -14,7 +14,6 @@ from mailbox import Message
 from smtplib import SMTPException
 
 import logging
-
 import sqlalchemy.exc
 
 from .. import db
@@ -27,63 +26,6 @@ from .BoardModels import Board, BoardForm
 ENTRY_LIMIT = 200
 logger = logging.getLogger(__name__)
 
-''' class Student(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(80), nullable=False)
-    course = db.Column(db.String(80), nullable=False)
-    grade = db.Column(db.String(10), nullable=False)
-    timestamp = db.Column(db.DateTime, default=datetime.utcnow())
-
-    course_name = lambda x : course_code[x]
-    course_code = {
-        'eng': 'English', 
-        'art': 'Art',
-        'mus': 'Music', 
-        'bus': 'Business',
-        'phy': 'Physic',
-        'his': 'History'
-                   }
-
-    def __repr__(self):
-         return f"<Message(name='{self.name}', course='{self.course}', grade='{self.grade}')>"
-
-    def table_exists(table_name):
-        from sqlalchemy import inspect
-        inspector = inspect(db.engine)
-        return inspector.has_table(table_name)
-
-    @student_bp.before_app_first_request
-    # logger.error('before_app_first_request method not found.')
-    def startup_check():
-        import os
-        if not table_exists('Message'):
-            current_app.logger.error("Table 'Message' does not exist. Creating table.")
-            db.create_all()  # This will create all tables defined by the models
-        else:
-            current_app.logger.debug("Table 'Message' exists.")
-
-    from functools import wraps
-    def roles_required(*roles):
-        def decorator(f):
-            @wraps(f)
-            def decorated_function(*args, **kwargs):
-                if current_user.role not in roles:
-                    flash('You do not have permission to access this page.', 'danger')
-                    return redirect(url_for('message_bp.show_message'))
-                return f(*args, **kwargs)
-            return decorated_function
-        return decorator
-
-    @Student_bp.route('/add', methods=['GET', 'POST'])
-    @login_required
-    @roles_required('teacher', 'TA')  # Allow both teachers and TAs to add message
-    def add_Message():
-        form = MessageForm()
-
-        return render_template('message/add_Message.html', form=form)
-
-
-'''
 
 
 @board_bp.route('/')     # D = Display
@@ -91,8 +33,8 @@ logger = logging.getLogger(__name__)
 def show_message():
     logger.debug('Contact Us-Show message route accessed.')
     form = BoardForm()
-    messages = Board.query.all()
-    return render_template('Board/board.html', messages=messages, form=form)
+    messages = Board.query.order_by(Board.timestamp.desc()).all()
+    return render_template('board.html', messages=messages, form=form)
 
 
 @board_bp.route('/general_add', methods=['GET', 'POST'])
@@ -115,23 +57,19 @@ def general_add_message():
             flash('Message added successfully!', 'success')
             return redirect(url_for('board_bp.general_add_message'))
 
-    except (SQLAlchemyError, IntegrityError, OperationalError) as e:
-        db.session.rollback()
-        error_message = SQL_exception(e)
-        flash (f'SQL commit error: {error_message}', 'error')
-        logger.error (f'SQL commit error: {error_message}') 
     except Exception as e:
-        flash (f'An unexpected error occurred: {e}', 'error')
-        logger.error(f'An unexpected error occurred: {e}')
+            error_message = handle_exception(e) 
+            flash (f'{error_message}', 'error')
+            logger.error(f'{error_message}')
 
 
-    return render_template('Board/board_general_add.html', form=sform)
+    return render_template('board_general_add.html', form=sform)
 
 @board_bp.route('/update/<int:id>', methods=['GET', 'POST'])
 @login_required
 @roles_required('sa', 'member')  # Only admins can register new users
 def reply_message(id):       # U = Update
-    current_app.logger.debug ('message-update route accessed.')
+    current_app.logger.debug ('message-reply route accessed.')
     stored_message = Board.query.get_or_404(id)
     sform = BoardForm(obj=stored_message)
 
@@ -140,6 +78,7 @@ def reply_message(id):       # U = Update
     
         if current_entries >= ENTRY_LIMIT:
             flash('The database has reached its limit of entries. Cannot add more message.', 'danger')
+            logger.error('The database has reached its limit of entries. Cannot add more message.')
             return redirect(url_for('board_bp.show_message'))
 
 
@@ -150,27 +89,16 @@ def reply_message(id):       # U = Update
                                     message=sform.message.data)
             db.session.add(replied_message)
             db.session.commit()
-            flash(f'client={name} message replied!', 'success')
+            flash(f'client={sform.name.data} message replied!', 'success')
+            logger.info(f'client={sform.name.data} message replied!')
             return redirect(url_for('board_bp.show_message'))
-
-    # except (SQLAlchemyError, IntegrityError, OperationalError) as e:
-    # # except sqlalchemy.exc as e:
-    #     db.session.rollback()
-    #     error_message = SQL_exception(e)
-    #     flash (f'SQL commit error: {error_message}', 'error')
-    #     logger.error (f'SQL commit error: {error_message}') 
-    # except Exception as e:
-    #     flash (f'An unexpected error occurred: {e}', 'error')
-    #     logger.error(f'An unexpected error occurred: {e}')
 
     except Exception as e:
             error_message = handle_exception(e) 
             flash (f'{error_message}', 'error')
             logger.error(f'{error_message}')
 
-
-    
-    return render_template('Board/board_reply.html', form=sform, message=stored_message)
+    return render_template('board_reply.html', form=sform, message=stored_message)
 
 @board_bp.route('/delete/<int:id>', methods=['POST'])
 @login_required
@@ -179,18 +107,20 @@ def delete_message(id):      # R = Remove
     
     try: 
         stored_message = Board.query.get_or_404(id)
-        db.session.delete(stored_message)
-        db.session.commit()
-
+        if stored_message.name == 'Admin'  :
+            raise ValueError('You cannot delete Admin message.')
+        else: 
+            db.session.delete(stored_message)
+            db.session.commit()
+            flash('Message deleted successfully!', 'success')
     except Exception as e:
         error_message = handle_exception(e) 
         flash (f'{error_message}', 'error')
         logger.error(f'{error_message}')
-
-    flash('Message deleted successfully!', 'success')
+    
     return redirect(url_for('board_bp.show_message'))
 
-""" no add message needed. 
+""" no add message route needed. 
 @board_bp.route('/add', methods=['GET', 'POST'])
 @login_required
 @roles_required('sa', 'member')
